Rename `initSharedState` to `storeCreator`.
