+++ src/lib/index.ts
@@ -1,6 +1,6 @@
 import { get, writable } from 'svelte/store';
 import type { Writable } from 'svelte/store';
-import { initSharedState, getActions } from './initSharedState';
+import { initSharedState, getActions } from './storeCreators';
 import withReduxDevtool from './middlewares/withReduxDevtools';
 import type { ExMiddleware } from './types/ExMiddleware';
 import type { ExSlice } from './types/ExSlice';
